+++ CHANGELOG.md
@@ -9,7 +9,12 @@
  _/\_/\_/\_,  ,_/_/\_/\_/\_/\_/\_/\_/\_/\_/\_
 ```
 
-## **dombili** (*v.0.6.1*)
+## **dombili** (*v0.7.0*)
+
+* Added a `parent()` method to get the parent elements of a node.
+* Added a `json()` method that assumes a **JSON** response from an **AJAX** request by default.
+
+## **dombili** (*v0.6.1*)
 
 * Minor updates and bugfixes.
 * Minor fixes in documentation.